Move broker into servicecatalog/client
